refactor: extract EventScheduleFields from CreateEditEventFields
Second event-form decomposition increment: move the entire date/time scheduling
section — single/multiple/recurring date modes, the pickers, all-day and
multi-day toggles, occurrence list, recurrence pattern + preview, and their
handlers — into a self-contained child component wired by formValues + an
updateFormValues event.

The one cross-section coupling (the parent's save-button validation read
start/end via the section's computeds) is removed by computing startBeforeEnd
inline from formValues, so the whole concern moves cleanly.

Parent drops to 472 lines (was 857 after the cover-image step; 1045 before
decomposition began). Existing CreateEditEventFields tests — including the
date-validation cases — still pass; adds focused tests for the new child.
